add jwt auth

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -47,18 +48,28 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path.replace('\\', '/');
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'Maximilian' },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
-      res.status(201).json({
-        message: 'Post created successfully!',
-        post: result,
+      User.findById(req.userId).then((user) => {
+        if (user) {
+          creator = user;
+          user.posts.push(post);
+          user.save().then(() => {
+            res.status(201).json({
+              message: 'Post created successfully!',
+              post: result,
+              creator: { _id: creator._id, name: creator.name },
+            });
+          });
+        }
       });
     })
     .catch((err) => {
@@ -114,6 +125,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Could not find post.');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -142,13 +158,24 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Could not find post.');
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
-      res.status(200).json({ message: 'Deleted post.' });
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      user.save().then(() => {
+        // console.log(result);
+        res.status(200).json({ message: 'Deleted post.' });
+      });
     })
     .catch((err) => {
       if (!err.statusCode) {
